+++ changelog.md
@@ -1,3 +1,17 @@
+## v1.5.1 [2019-03-12]
+
+#### Added
+
+* Advanced Configuration in North instance modal
+* Added build version info on settings page
+
+#### Fixed
+
+* Certificate store upload issue
+* Allow negative numbers to be entered in numeric fields
+* HTTP Error (4xx / 5xx) response handling
+* Days info in uptime
+
 ## v1.5.0 [2019-02-07]
 
 #### Added